Fix: exception 추가 IllegalArgumentException 수정

+++ src/main/java/sideproject/board/board/service/BoardService.java
@@ -1,5 +1,7 @@
 package sideproject.board.board.service;
 
+import static sideproject.board.board.exception.ExceptionCode.*;
+
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -9,6 +11,7 @@ import lombok.RequiredArgsConstructor;
 import sideproject.board.board.controller.dto.requests.UpdateRequest;
 import sideproject.board.board.domain.BoardRepository;
 import sideproject.board.board.domain.entity.Board;
+import sideproject.board.board.exception.BadRequestException;
 
 @Service
 @RequiredArgsConstructor
@@ -32,7 +35,7 @@ public class BoardService {
 	public Board getOneBoard(Long id) {
 
 		return boardRepository.findById(id)
-			.orElseThrow(() -> new IllegalArgumentException("아이디 찾을 수 없습니다."));
+			.orElseThrow(() -> new BadRequestException(NOT_FOUND_BOARD_ID));
 
 	}
 
@@ -40,7 +43,7 @@ public class BoardService {
 	public Board updateBoard(Long id, UpdateRequest request) {
 
 		Board board = boardRepository.findById(id)
-			.orElseThrow(() -> new IllegalArgumentException("아이디 찾을 수 없습니다."));
+			.orElseThrow(() -> new BadRequestException(NOT_FOUND_BOARD_ID));
 
 		board.update(id, request.getTitle(), request.getContent(), request.getPrice());
 
@@ -49,6 +52,9 @@ public class BoardService {
 
 	@Transactional
 	public void deleteBoard(Long id) {
+		if (!boardRepository.existsById(id)) {
+			throw new BadRequestException(NOT_FOUND_BOARD_ID);
+		}
 		boardRepository.deleteById(id);
 	}
 }
